chore(desktop): split engine-event tests under file-size cap
The sendPrompt arg-count fix pushed engine-control-plane.test.ts to 606
lines, two over the 600 cap, failing the file-size CI gate. Per repo
policy, split rather than strip comments.

Extract the 'engine event handling' describe block (9 tests) into a new
sibling engine-control-plane-engine-events.test.ts, mirroring the
established engine-control-plane-*.test.ts split pattern. The new file
carries its own hoisted vi.mock setup (vitest hoists per file, so the
mocks cannot be shared via import). The original drops to 336 lines and
retains session-lifecycle / prompt-routing / remote-validation /
reconnect coverage; its top doc comment now points at the sibling.

All 26 control-plane tests pass across the three files (12 + 9 + 5);
the pre-existing engine-control-plane-events.test.ts (issue #230 B1
guard, 5 tests) is untouched. Both files are under the cap.
