+++ desktop/src/main/__tests__/engine-control-plane.test.ts
@@ -1,7 +1,8 @@
 /**
  * EngineControlPlane Tests
  *
- * Validates session lifecycle, prompt routing, and engine event handling.
+ * Validates session lifecycle and prompt routing. Engine event handling
+ * lives in the sibling engine-control-plane-engine-events.test.ts.
  * Covers three shipped bugs: engine_dead exit code 0 treated as error,
  * conversationId never set, sessionId not passed through EngineConfig.
  */
@@ -278,277 +279,6 @@ describe('EngineControlPlane', () => {
         undefined,
         undefined,
       )
-    })
-  })
-
-  describe('engine event handling', () => {
-    it('engine_dead with code 0 sets status idle without error', async () => {
-      const tabId = cp.createTab()
-      await cp.submitPrompt(tabId, 'req-1', makeRunOptions())
-
-      const errors: any[] = []
-      cp.on('error', (_tid: string, err: any) => errors.push(err))
-
-      const statusChanges: any[] = []
-      cp.on('tab-status-change', (tid: string, newS: string, oldS: string) => {
-        statusChanges.push({ tid, newS, oldS })
-      })
-
-      // Simulate engine_dead with exit code 0
-      expect(capturedEventHandler).not.toBeNull()
-      capturedEventHandler!(tabId, { type: 'engine_dead', exitCode: 0 })
-
-      expect(errors).toHaveLength(0)
-      const tab = cp.getTabStatus(tabId)
-      expect(tab?.status).toBe('idle')
-    })
-
-    it('engine_dead with code 1 emits error and sets status dead', async () => {
-      const tabId = cp.createTab()
-      await cp.submitPrompt(tabId, 'req-1', makeRunOptions())
-
-      const errors: any[] = []
-      cp.on('error', (_tid: string, err: any) => errors.push(err))
-
-      capturedEventHandler!(tabId, { type: 'engine_dead', exitCode: 1, stderrTail: ['segfault'] })
-
-      expect(errors).toHaveLength(1)
-      expect(errors[0].exitCode).toBe(1)
-      expect(errors[0].message).toContain('code 1')
-      const tab = cp.getTabStatus(tabId)
-      expect(tab?.status).toBe('dead')
-    })
-
-    it('engine_status idle emits task_complete', async () => {
-      const tabId = cp.createTab()
-      await cp.submitPrompt(tabId, 'req-1', makeRunOptions())
-
-      const events: any[] = []
-      cp.on('event', (tid: string, ev: any) => events.push({ tid, ev }))
-
-      capturedEventHandler!(tabId, {
-        type: 'engine_status',
-        fields: { state: 'idle', totalCostUsd: 0.01 },
-      })
-
-      const taskComplete = events.find((e) => e.ev.type === 'task_complete')
-      expect(taskComplete).toBeDefined()
-      expect(taskComplete.ev.costUsd).toBe(0.01)
-
-      const tab = cp.getTabStatus(tabId)
-      expect(tab?.status).toBe('idle')
-    })
-
-    it('engine_status idle with AskUserQuestion denial emits task_complete with permissionDenials and sets status completed', async () => {
-      const tabId = cp.createTab()
-      await cp.submitPrompt(tabId, 'req-1', makeRunOptions())
-
-      const events: any[] = []
-      cp.on('event', (tid: string, ev: any) => events.push({ tid, ev }))
-
-      const statusChanges: any[] = []
-      cp.on('tab-status-change', (tid: string, newS: string, oldS: string) => {
-        statusChanges.push({ tid, newS, oldS })
-      })
-
-      capturedEventHandler!(tabId, {
-        type: 'engine_status',
-        fields: {
-          state: 'idle',
-          totalCostUsd: 0.02,
-          permissionDenials: [
-            { toolName: 'AskUserQuestion', toolUseId: 'ask-1', toolInput: { question: 'Pick one' } },
-          ],
-        },
-      })
-
-      // task_complete carries the denials
-      const taskComplete = events.find((e) => e.ev.type === 'task_complete')
-      expect(taskComplete).toBeDefined()
-      expect(taskComplete.ev.permissionDenials).toHaveLength(1)
-      expect(taskComplete.ev.permissionDenials[0].toolName).toBe('AskUserQuestion')
-
-      // Control plane status is 'completed' (not 'idle')
-      const tab = cp.getTabStatus(tabId)
-      expect(tab?.status).toBe('completed')
-
-      // tab-status-change emitted as 'completed'
-      const completedChange = statusChanges.find((s) => s.newS === 'completed')
-      expect(completedChange).toBeDefined()
-    })
-
-    it('engine_status idle with AskUserQuestion denial skips duplicate idle', async () => {
-      const tabId = cp.createTab()
-      await cp.submitPrompt(tabId, 'req-1', makeRunOptions())
-
-      const events: any[] = []
-      cp.on('event', (tid: string, ev: any) => events.push({ tid, ev }))
-
-      // First idle with denial → completed
-      capturedEventHandler!(tabId, {
-        type: 'engine_status',
-        fields: {
-          state: 'idle',
-          permissionDenials: [
-            { toolName: 'AskUserQuestion', toolUseId: 'ask-1', toolInput: { question: 'Yes?' } },
-          ],
-        },
-      })
-      expect(cp.getTabStatus(tabId)?.status).toBe('completed')
-
-      // Second idle (cost-only update) should be skipped by the completed guard
-      const eventsBefore = events.length
-      capturedEventHandler!(tabId, {
-        type: 'engine_status',
-        fields: { state: 'idle', totalCostUsd: 0.05 },
-      })
-
-      // No additional task_complete should have been emitted
-      const taskCompletes = events.filter((e) => e.ev.type === 'task_complete')
-      expect(taskCompletes).toHaveLength(1) // only the first one
-
-      // Status remains completed
-      expect(cp.getTabStatus(tabId)?.status).toBe('completed')
-    })
-
-    it('engine_tool_update emits tool_call_update with partialInput', async () => {
-      const tabId = cp.createTab()
-      await cp.submitPrompt(tabId, 'req-1', makeRunOptions())
-
-      const events: any[] = []
-      cp.on('event', (tid: string, ev: any) => events.push({ tid, ev }))
-
-      capturedEventHandler!(tabId, {
-        type: 'engine_tool_update',
-        toolId: 'tool-42',
-        partialInput: '{"file_path":"/tmp',
-      })
-
-      const update = events.find((e) => e.ev.type === 'tool_call_update')
-      expect(update).toBeDefined()
-      expect(update.ev.toolId).toBe('tool-42')
-      expect(update.ev.partialInput).toBe('{"file_path":"/tmp')
-    })
-
-    it('engine_tool_complete emits tool_call_complete with index', async () => {
-      const tabId = cp.createTab()
-      await cp.submitPrompt(tabId, 'req-1', makeRunOptions())
-
-      const events: any[] = []
-      cp.on('event', (tid: string, ev: any) => events.push({ tid, ev }))
-
-      capturedEventHandler!(tabId, {
-        type: 'engine_tool_complete',
-        index: 3,
-      })
-
-      const complete = events.find((e) => e.ev.type === 'tool_call_complete')
-      expect(complete).toBeDefined()
-      expect(complete.ev.index).toBe(3)
-    })
-
-    it('stale engine_status idle after resetTabSession does not synthesize task_complete', async () => {
-      const tabId = cp.createTab()
-
-      // Step 1: Submit first prompt → tab becomes running
-      await cp.submitPrompt(tabId, 'req-1', makeRunOptions())
-      expect(cp.getTabStatus(tabId)?.status).toBe('running')
-
-      // Step 2: First session completes with ExitPlanMode denial → tab becomes completed
-      const events: any[] = []
-      cp.on('event', (tid: string, ev: any) => events.push({ tid, ev }))
-
-      capturedEventHandler!(tabId, {
-        type: 'engine_status',
-        fields: {
-          state: 'idle',
-          totalCostUsd: 0.01,
-          permissionDenials: [
-            { toolName: 'ExitPlanMode', toolUseId: 'exit-1', toolInput: {} },
-          ],
-        },
-      })
-      expect(cp.getTabStatus(tabId)?.status).toBe('completed')
-      expect(events.filter((e) => e.ev.type === 'task_complete')).toHaveLength(1)
-
-      // Step 3: resetTabSession (simulates onImplement flow) → tab resets to idle
-      cp.resetTabSession(tabId)
-      expect(cp.getTabStatus(tabId)?.status).toBe('idle')
-      expect(cp.getTabStatus(tabId)?.activeRequestId).toBeNull()
-
-      // Step 4: Submit new prompt (implementation) → tab becomes running again
-      await cp.submitPrompt(tabId, 'req-2', makeRunOptions({ prompt: 'implement' }))
-      expect(cp.getTabStatus(tabId)?.status).toBe('running')
-
-      // Step 5: Stale engine_status idle from the OLD dying session arrives
-      const eventsBefore = events.length
-      capturedEventHandler!(tabId, {
-        type: 'engine_status',
-        fields: { state: 'idle' },
-      })
-
-      // No additional task_complete should have been emitted — the stale idle
-      // is rejected because tab.status was 'idle' at the time resetTabSession
-      // cleared it, and then submitPrompt set it to 'running'. However, the
-      // critical path is: after resetTabSession sets status='idle', a stale
-      // idle arriving BEFORE submitPrompt runs would also be caught.
-      // Either way, only the original task_complete from step 2 should exist.
-      //
-      // In this test the stale idle arrives AFTER submitPrompt (status='running'),
-      // so it WOULD synthesize a spurious task_complete without the fix.
-      // With the fix, the idle guard at tab.status === 'idle' catches it during
-      // the window between resetTabSession and submitPrompt. Since submitPrompt
-      // already ran here, we verify the broader scenario: the idle event does
-      // produce a task_complete (genuine completion of req-2). The real
-      // protection for the race is tested in the next assertion block.
-
-      // Verify the stale idle scenario: resetTabSession → stale idle (no submitPrompt yet)
-      // This is the exact race the fix targets.
-    })
-
-    it('stale engine_status idle after resetTabSession but before new submitPrompt is rejected', async () => {
-      const tabId = cp.createTab()
-
-      // Step 1: Submit prompt → running
-      await cp.submitPrompt(tabId, 'req-1', makeRunOptions())
-      expect(cp.getTabStatus(tabId)?.status).toBe('running')
-
-      const events: any[] = []
-      cp.on('event', (tid: string, ev: any) => events.push({ tid, ev }))
-
-      // Step 2: Session completes with ExitPlanMode → completed
-      capturedEventHandler!(tabId, {
-        type: 'engine_status',
-        fields: {
-          state: 'idle',
-          permissionDenials: [
-            { toolName: 'ExitPlanMode', toolUseId: 'exit-1', toolInput: {} },
-          ],
-        },
-      })
-      expect(cp.getTabStatus(tabId)?.status).toBe('completed')
-      const firstTaskCompletes = events.filter((e) => e.ev.type === 'task_complete')
-      expect(firstTaskCompletes).toHaveLength(1)
-
-      // Step 3: resetTabSession → status becomes 'idle'
-      cp.resetTabSession(tabId)
-      expect(cp.getTabStatus(tabId)?.status).toBe('idle')
-
-      // Step 4: Stale idle from dying old session arrives BEFORE new submitPrompt
-      // This is the exact race condition — without the fix, status is 'idle'
-      // which is not 'completed', so the old guard would let it through and
-      // synthesize a spurious task_complete.
-      capturedEventHandler!(tabId, {
-        type: 'engine_status',
-        fields: { state: 'idle' },
-      })
-
-      // With the fix, no additional task_complete is synthesized
-      const allTaskCompletes = events.filter((e) => e.ev.type === 'task_complete')
-      expect(allTaskCompletes).toHaveLength(1) // still just the one from step 2
-
-      // Status remains idle
-      expect(cp.getTabStatus(tabId)?.status).toBe('idle')
     })
   })
